requirements.txt

tensorflow
gradio
Pillow
joblib
scikit-learn
numpy
